[fix][build] Publish the parent POM (and all modules) to the same remote repository
### Motivation

The root project publishes `org.apache.pulsar:pulsar` as a POM-only parent
artifact that every child POM references via `<parent>`. Its publishing was
hand-rolled in the root build script and declared only the local `localDeploy`
repository — never the ASF Nexus `apacheReleases` / `apacheSnapshots` repos that
`pulsar.publish-conventions` configures for the modules. As a result the root had
no `publish…ToApache…Repository` task, so a `publishAllPublicationsToApache…`
run uploaded every module but silently skipped the parent POM, leaving the
published modules unresolvable (missing parent).

### Modifications

- New convention `pulsar.publish-repositories-conventions`: owns the publish
  TARGETS shared by every published project — the `localDeploy` + ASF Nexus
  repositories, the snapshot/release validation, the `mavenPublishLock` upload
  serialization service, and GPG signing. It creates no publication and does not
  touch the `base`/`java` plugins, so it applies equally to the POM-only root
  project and to every java-library / java-platform module.
- `pulsar.publish-conventions` now applies it and keeps only the per-module
  publication wiring (component, POM metadata, sources/javadoc, artifactId).
- The root build script applies it (replacing its bare `maven-publish` + `signing`
  and its localDeploy-only repositories block), so the parent POM now uploads to
  the same Nexus staging repository as the modules, through the same lock.
- The snapshot/release validation now matches on the publish task name
  (`…ToApacheSnapshotsRepository`) instead of reading `task.repository`, which is
  null while the root's POM-only publish task is being created.

Verified: the root `publishMavenPublicationToApacheSnapshotsRepository` task now
exists and uploads `org/apache/pulsar/pulsar/.../pulsar-*.pom` to a file:// remote
alongside the modules; configuration cache stores and reuses without problems.
